+++ aofl-js-packages/form-validate/modules/compare.js
@@ -1,27 +1,30 @@
 /**
- * @summary is-equal
+ * @summary compare
  * @version 3.0.0
  * @since 1.0.0
  * @author Arian Khosravi <arian.khosravi@aofl.com>
+ * @author Isaac Yoon <isaac.yoon@aofl.com>
  */
-import {compare} from './compare';
+import {get} from '@aofl/object-utils';
 /**
  * Tracks two properties of any class to be equal.
  *
  * @memberof module:@aofl/form-validate
  *
  * @param {String} propPath
+ * @param {Function} comparator
  *
  * @return {Function} True or false if the given value passes the
  *                   given regex test is valid.
  */
-const isEqual = (propPath) => {
-  return compare(propPath, (value, otherValue) => {
+const compare = (propPath, comparator) => {
+  return /** @this HTMLElement */ function compareValidator(value) {
+    const otherVal = get(this.target, propPath);
     if (typeof otherVal === 'undefined') { return false; }
-    return value === otherValue;
-  });
+    return comparator(value, otherVal);
+  };
 };
 
 export {
-  isEqual
+  compare
 };