+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased](https://github.com/AgeOfLearning/aofl/compare/v3.2.1...development) - Date
+
+### [Added]
+- @aofl/form-validate - exports ValidationFunction and ValidationProperty classes
+- @aofl/form-validate - compare validator
+- @aofl/form-validate - pattern validator
+### [Changed]
+### [Deprecated]
+### [Removed]
+### [Fixed]
+### [Security]
+
+---
 ## [3.2.1](https://github.com/AgeOfLearning/aofl/compare/v3.2.0...v3.2.1) - 2019-11-23
 
 ### [Added]