+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - @aofl/form-validate - exports ValidationFunction and ValidationProperty classes
 - @aofl/form-validate - compare validator
 - @aofl/form-validate - pattern validator
+- @aofl/cli-lib - added npm run
 ### [Changed]
 - @aofl/picture - img inherits height, width, max-height, max-width from aofl-img
 - @aofl/picture - Uses naturalWidth and naturalHeight of the image